Add files via upload
The aim of this project is to develop a CNN-based model that can accurately predict the presence of diseases in cotton plants based on images of their leaves. The model should be capable of distinguishing between healthy cotton plants and those affected by various diseases, such as cotton leaf spot, cotton boll rot, and cotton leaf curl virus, among others.